requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
SQLAlchemy
jinja2
pillow
python-multipart
requests
httpx
edge-tts
deep-translator
langgraph
langchain-openai
langchain-google-genai
langchain-anthropic
langchain-ollama
playwright
gunicorn
